+++ Changelog.md
@@ -1,6 +1,7 @@
 Unreleased
 ---------------------------
 
+* [Enhancement] Create admin page for stacks
 * [Bug fix] Avoid unhandled exception on SSH close
 
 Version 3.3.0 (2019-10-15)